backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
meilisearch
python-dotenv
hashlib